debugged getConversion/ all testcases passed

+++ app/src/main/java/currency/converter/App.java
@@ -40,7 +40,7 @@ public class App {
                         String currCurrency = input_list[1];
                         String newCurrency = input_list[2];
                         float amount = Float.parseFloat(input_list[3]);
-                        float currency = handler.convertCurrency(amount, currCurrency, newCurrency);
+                        double currency = handler.convertCurrency(amount, currCurrency, newCurrency);
                         System.out.println(
                                 "The conversion of " + amount + currCurrency + " is: " + currency + " " + newCurrency);
                     } else {
